refactor: move auth config to config folder
