--- lib/server/auth/auth.config.ts
@@ -1,5 +1,0 @@
-export const inactivityTimeoutSeconds = 60 * 60 * 24 * 7; /** 7 days. */
-export const activityCheckIntervalSeconds = 60 * 60; /** 1 hour. */
-
-export const sessionCookieName = "session";
-export const sessionCookieMaxAgeSeconds = 60 * 60 * 24 * 30; /** 30 days. */
